Widget JS Error
JS: After we're installing and then deactivating the WPML plugin and added a language switcher to the sidebar, we couldn't customize our theme, and we couldn't save the modifications of widgets on Widgets screen.

When I opened the Console, I saw JS errors, so I started to debugging the code, and I found out, that sometimes the value of this.setting.get is not an array, just null, and the value of post.blocks variable is undefined.